+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Changelog
 * Deprecate :meth:`pyvep.filter_lof/clinsig` methods.
 * Update :meth:`pymaf.MafFrame.from_vcf` method to extract genotype keys (e.g. DP, AD, AF).
 * Update :command:`bam_slice` and :command:`bam_rename` commands.
+* Deprecate :meth:`pybam.rename`.
 
 0.14.0 (2021-06-20)
 -------------------